requirements.txt
flask==3.0.3
google-generativeai==0.7.2
google-api-python-client==2.146.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
pypdf==4.3.1
streamlit==1.28.0